backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.19.2
numpy==1.26.4
pillow==10.4.0
python-multipart==0.0.12
slowapi==0.1.9
